Move PokemonMove to its own file
