+++ src/data/mystery-encounters/encounters/bug-type-superfan-encounter.ts
@@ -24,7 +24,7 @@ import { TrainerType } from "#enums/trainer-type";
 import { Species } from "#enums/species";
 import type { PlayerPokemon } from "#app/field/pokemon";
 import type Pokemon from "#app/field/pokemon";
-import { PokemonMove } from "#app/field/pokemon";
+import { PokemonMove } from "#app/data/moves/pokemon-move";
 import { getEncounterText, showEncounterDialogue } from "#app/data/mystery-encounters/utils/encounter-dialogue-utils";
 import { LearnMovePhase } from "#app/phases/learn-move-phase";
 import { Moves } from "#enums/moves";
