+++ src/data/mystery-encounters/encounters/fiery-fallout-encounter.ts
@@ -26,7 +26,7 @@ import { Gender } from "#app/data/gender";
 import { PokemonType } from "#enums/pokemon-type";
 import { BattlerIndex } from "#app/battle";
 import type Pokemon from "#app/field/pokemon";
-import { PokemonMove } from "#app/field/pokemon";
+import { PokemonMove } from "#app/data/moves/pokemon-move";
 import { Moves } from "#enums/moves";
 import { EncounterBattleAnim } from "#app/data/battle-anims";
 import { WeatherType } from "#enums/weather-type";
