+++ src/system/pokemon-data.ts
@@ -5,7 +5,8 @@ import type { Nature } from "#enums/nature";
 import type { PokeballType } from "#enums/pokeball";
 import { getPokemonSpecies, getPokemonSpeciesForm } from "../data/pokemon-species";
 import { Status } from "../data/status-effect";
-import Pokemon, { EnemyPokemon, PokemonMove, PokemonSummonData } from "../field/pokemon";
+import Pokemon, { EnemyPokemon, PokemonSummonData } from "../field/pokemon";
+import { PokemonMove } from "#app/data/moves/pokemon-move";
 import { TrainerSlot } from "#enums/trainer-slot";
 import type { Variant } from "#app/sprites/variant";
 import { loadBattlerTag } from "../data/battler-tags";
