+++ test/imports.test.ts
@@ -4,7 +4,7 @@ import { describe, expect, it } from "vitest";
 async function importModule() {
   try {
     initStatsKeys();
-    const { PokemonMove } = await import("#app/field/pokemon");
+    const { PokemonMove } = await import("#app/data/moves/pokemon-move");
     const { Species } = await import("#enums/species");
     return {
       PokemonMove,
