+++ test/moves/sketch.test.ts
@@ -1,7 +1,8 @@
 import { Abilities } from "#enums/abilities";
 import { Moves } from "#enums/moves";
 import { Species } from "#enums/species";
-import { MoveResult, PokemonMove } from "#app/field/pokemon";
+import { MoveResult } from "#app/field/pokemon";
+import { PokemonMove } from "#app/data/moves/pokemon-move";
 import GameManager from "#test/testUtils/gameManager";
 import Phaser from "phaser";
 import { afterEach, beforeAll, beforeEach, describe, expect, it, vi } from "vitest";
