+++ test/mystery-encounter/encounters/clowning-around-encounter.test.ts
@@ -15,7 +15,7 @@ import {
 import { Moves } from "#enums/moves";
 import type BattleScene from "#app/battle-scene";
 import type Pokemon from "#app/field/pokemon";
-import { PokemonMove } from "#app/field/pokemon";
+import { PokemonMove } from "#app/data/moves/pokemon-move";
 import { Mode } from "#app/ui/ui";
 import { MysteryEncounterOptionMode } from "#enums/mystery-encounter-option-mode";
 import { MysteryEncounterTier } from "#enums/mystery-encounter-tier";
